+++ misc.py
@@ -1,58 +1,62 @@
 from datetime import datetime
 
-from notifier import SlackColourThresholds
+from constants import SlackColourThresholds
 from stats import TimeIntervalData
 from coinbase import Currency
 from history import History
 
-def should_post(history: History, stats: TimeIntervalData, threshold: float):
-    if history.rising != stats.is_diff_positive:
-        print(f"Last change was in the opposite direction")
-        return True
+class Misc:
+    @staticmethod
+    def should_post(history: History, stats: TimeIntervalData, threshold: float):
+        if history.rising != stats.is_diff_positive:
+            print(f"Last change was in the opposite direction")
+            return True
 
-    if history.ema_reset:
-        return True
-    else:
-        # Allow if increase is greater again
-        print("Price has not gone back within the EMA threshold since the last post")
-        if history.rising:
-            required_perc_diff = (1 + threshold / 100)
-            threshold_sign_str = "above"
-        else:
-            required_perc_diff = (1 - threshold / 100)
-            threshold_sign_str = "below"
-
-        new_threshold = history.price * required_perc_diff
-        print(f"To post again the current price must be {threshold_sign_str}: {new_threshold:.0f}")
-        if (history.rising and stats.cur_price > new_threshold) or \
-                (not history.rising and stats.cur_price < new_threshold):
-            print(f"Beats new threshold price ({stats.cur_price:.0f}/{new_threshold:.0f})")
+        if history.ema_reset:
             return True
         else:
-            print(f"Does not beat new threshold price: ({stats.cur_price:.0f}/{new_threshold:.0f})")
-            return False
+            # Allow if increase is greater again
+            print("Price has not gone back within the EMA threshold since the last post")
+            if history.rising:
+                required_perc_diff = (1 + threshold / 100)
+                threshold_sign_str = "above"
+            else:
+                required_perc_diff = (1 - threshold / 100)
+                threshold_sign_str = "below"
 
-def format_stat(stat: TimeIntervalData, stats: TimeIntervalData, text_pretext: str, pretext=None):
-    diff = stats.cur_price - stat.cur_price
-    diff /= stat.cur_price
-    diff *= 100
+            new_threshold = history.price * required_perc_diff
+            print(f"To post again the current price must be {threshold_sign_str}: {new_threshold:.0f}")
+            if (history.rising and stats.cur_price > new_threshold) or \
+                    (not history.rising and stats.cur_price < new_threshold):
+                print(f"Beats new threshold price ({stats.cur_price:.0f}/{new_threshold:.0f})")
+                return True
+            else:
+                print(f"Does not beat new threshold price: ({stats.cur_price:.0f}/{new_threshold:.0f})")
+                return False
 
-    if diff > SlackColourThresholds.GOOD:
-        colour = "good"
-    elif diff > SlackColourThresholds.NEUTRAL:
-        colour = ""
-    elif diff > SlackColourThresholds.WARNING:
-        colour = "warning"
-    else:
-        colour = "danger"
+    @staticmethod
+    def format_stat(stat: TimeIntervalData, stats: TimeIntervalData, text_pretext: str, pretext=None):
+        diff = stats.cur_price - stat.cur_price
+        diff /= stat.cur_price
+        diff *= 100
 
-    text = f"{text_pretext}{Currency.SECONDARY_CURRENCY_SYMBOL}{stat.cur_price:,.0f} ({diff:+.2f}%)"
-    attachment = {"fallback": "some price changes", "text": text, "color": colour}
-    if pretext is not None:
-        attachment['pretext'] = pretext
+        if diff > SlackColourThresholds.GOOD:
+            colour = "good"
+        elif diff > SlackColourThresholds.NEUTRAL:
+            colour = ""
+        elif diff > SlackColourThresholds.WARNING:
+            colour = "warning"
+        else:
+            colour = "danger"
 
-    return attachment
+        text = f"{text_pretext}{Currency.SECONDARY_CURRENCY_SYMBOL}{stat.cur_price:,.0f} ({diff:+.2f}%)"
+        attachment = {"fallback": "some price changes", "text": text, "color": colour}
+        if pretext is not None:
+            attachment['pretext'] = pretext
 
-def cur_time_hours():
-    cur_time = datetime.utcnow()
-    return cur_time.replace(minute=0, second=0, microsecond=0)
+        return attachment
+
+    @staticmethod
+    def cur_time_hours():
+        cur_time = datetime.utcnow()
+        return cur_time.replace(minute=0, second=0, microsecond=0)
